create module for DatabaseCommit benchmark
addresses review comment https://github.com/NomicFoundation/hardhat/pull/3953#discussion_r1231560608
